Reorganize ensemble modules into ensemble package
