+++ ert_shared/ensemble_evaluator/ensemble/base.py
