+++ tests/ensemble_evaluator/ensemble_evaluator_utils.py
@@ -6,8 +6,8 @@ from cloudevents.http import CloudEvent, to_json
 
 from ert_shared.ensemble_evaluator.client import Client
 from ert_shared.ensemble_evaluator.entity import identifiers as identifiers
-from ert_shared.ensemble_evaluator.entity.ensemble import _BaseJob, _Realization, _Step
-from ert_shared.ensemble_evaluator.entity.ensemble_base import _Ensemble
+from ert_shared.ensemble_evaluator.ensemble.builder import _BaseJob, _Realization, _Step
+from ert_shared.ensemble_evaluator.ensemble.base import _Ensemble
 
 
 def _mock_ws(host, port, messages, delay_startup=0):
