+++ tests/ensemble_evaluator/test_ensemble_builder.py
@@ -4,7 +4,7 @@ from unittest.mock import MagicMock, Mock
 import pytest
 from attr.validators import matches_re
 
-import ert_shared.ensemble_evaluator.entity.ensemble as ee
+import ert_shared.ensemble_evaluator.ensemble.builder as ee
 
 
 def test_build_ensemble():
