+++ tests/ensemble_evaluator/test_entity.py
@@ -8,7 +8,7 @@ import ert_shared.status.entity.state as state
 from ert_shared.ensemble_evaluator.entity import command
 from ert_shared.ensemble_evaluator.entity import identifiers as ids
 from ert_shared.ensemble_evaluator.entity import tool
-from ert_shared.ensemble_evaluator.entity.ensemble_base import _EnsembleStateTracker
+from ert_shared.ensemble_evaluator.ensemble.base import _EnsembleStateTracker
 from ert_shared.ensemble_evaluator.entity.snapshot import (
     Job,
     PartialSnapshot,
